requirements.txt
certifi==2020.12.5
chardet==4.0.0
Django==2.2
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
idna==2.10
importlib-metadata==3.4.0
Markdown==3.3.3
python-decouple==3.4
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.3
zipp==3.4.0
